+++ src/server/prod.ts
@@ -5,7 +5,7 @@ import * as http from 'http'
 import * as minimist from 'minimist'
 import * as favicon from 'serve-favicon'
 import * as sio from 'socket.io'
-import { RobotSimulator } from '../simulators/robot_simulator'
+import { VirtualRobots } from '../simulators/virtual_robots'
 import { SensorDataSimulator } from '../simulators/sensor_data_simulator'
 import { WebSocketProxyNUClearNetServer } from './nuclearnet/web_socket_proxy_nuclearnet_server'
 import { WebSocketServer } from './nuclearnet/web_socket_server'
@@ -30,7 +30,7 @@ server.listen(port, () => {
 })
 
 if (withSimulators) {
-  const robotSimulator = RobotSimulator.of({
+  const robotSimulator = VirtualRobots.of({
     fakeNetworking: true,
     numRobots: 1,
     simulators: [
