+++ src/simulators/virtual_robot.ts
@@ -1,48 +1,17 @@
-import { Clock } from '../../src/server/time/clock'
-import { DirectNUClearNetClient } from '../server/nuclearnet/direct_nuclearnet_client'
-import { FakeNUClearNetClient } from '../server/nuclearnet/fake_nuclearnet_client'
-import { NodeSystemClock } from '../server/time/node_clock'
-import { NUClearNetClient } from '../shared/nuclearnet/nuclearnet_client'
-import { flatMap } from './flat_map'
 import { Simulator } from './simulator'
-
-export class RobotSimulator {
-  private robots: SimulatedRobot[]
-
-  public constructor(opts: { robots: SimulatedRobot[] }) {
-    this.robots = opts.robots
-  }
-
-  public static of(opts: { fakeNetworking: boolean, numRobots: number; simulators: Simulator[] }): RobotSimulator {
-    const robots = range(opts.numRobots).map(index => SimulatedRobot.of({
-      fakeNetworking: opts.fakeNetworking,
-      name: `Simulated Robot #${index + 1}`,
-      simulators: opts.simulators,
-    }))
-    return new RobotSimulator({ robots })
-  }
-
-  public simulateWithFrequency(frequency: number): () => void {
-    const stops = this.robots.map((robot, index) => robot.simulateWithFrequency(frequency, index, this.robots.length))
-    return () => stops.forEach(stop => stop())
-  }
-
-  public simulate(): void {
-    this.robots.forEach((robot, index) => robot.simulate(index, this.robots.length))
-  }
-
-  public connect(): void {
-    this.robots.forEach(robot => robot.connect())
-  }
-}
-
+import { NUClearNetClient } from '../shared/nuclearnet/nuclearnet_client'
+import { Clock } from '../server/time/clock'
+import { FakeNUClearNetClient } from '../server/nuclearnet/fake_nuclearnet_client'
+import { DirectNUClearNetClient } from '../server/nuclearnet/direct_nuclearnet_client'
+import { NodeSystemClock } from '../server/time/node_clock'
+import { flatMap } from './flat_map'
 type SimulatedRobotOpts = {
   fakeNetworking: boolean
   name: string
   simulators: Simulator[]
 }
 
-export class SimulatedRobot {
+export class VirtualRobot {
   private name: string
   private simulators: Simulator[]
 
@@ -53,10 +22,10 @@ export class SimulatedRobot {
     this.simulators = opts.simulators
   }
 
-  public static of(opts: SimulatedRobotOpts): SimulatedRobot {
+  public static of(opts: SimulatedRobotOpts): VirtualRobot {
     const network = opts.fakeNetworking ? FakeNUClearNetClient.of() : DirectNUClearNetClient.of()
     const clock = NodeSystemClock
-    return new SimulatedRobot(network, clock, opts)
+    return new VirtualRobot(network, clock, opts)
   }
 
   public simulateWithFrequency(frequency: number, index: number, numRobots: number) {
@@ -92,12 +61,4 @@ export class SimulatedRobot {
   public connect(): () => void {
     return this.network.connect({ name: this.name })
   }
-}
-
-function range(n: number): number[] {
-  const arr = []
-  for (let i = 0; i < n; i++) {
-    arr.push(i)
-  }
-  return arr
 }
