+++ src/tests/networking_integration.tests.ts
@@ -5,8 +5,8 @@ import { FakeNUClearNetClient } from '../server/nuclearnet/fake_nuclearnet_clien
 import { FakeNUClearNetServer } from '../server/nuclearnet/fake_nuclearnet_server'
 import { NodeSystemClock } from '../server/time/node_clock'
 import { message } from '../shared/proto/messages'
-import { RobotSimulator } from '../simulators/robot_simulator'
-import { SimulatedRobot } from '../simulators/robot_simulator'
+import { VirtualRobots } from '../simulators/virtual_robots'
+import { VirtualRobot } from '../simulators/virtual_robot'
 import { SensorDataSimulator } from '../simulators/sensor_data_simulator'
 import Sensors = message.input.Sensors
 import VisionObject = message.vision.VisionObject
@@ -15,7 +15,7 @@ import Overview = message.support.nubugger.Overview
 describe('Networking Integration', () => {
   let nuclearnetServer: FakeNUClearNetServer
   let nusightNetwork: NUsightNetwork
-  let robotSimulator: RobotSimulator
+  let virtualRobots: VirtualRobots
   let disconnectNusightNetwork: () => void
 
   beforeEach(() => {
@@ -23,9 +23,9 @@ describe('Networking Integration', () => {
     nusightNetwork = createNUsightNetwork()
     disconnectNusightNetwork = nusightNetwork.connect({ name: 'nusight' })
 
-    robotSimulator = new RobotSimulator({
+    virtualRobots = new VirtualRobots({
       robots: [
-        new SimulatedRobot(
+        new VirtualRobot(
           new FakeNUClearNetClient(nuclearnetServer),
           NodeSystemClock,
           {
@@ -38,7 +38,7 @@ describe('Networking Integration', () => {
         ),
       ],
     })
-    robotSimulator.connect()
+    virtualRobots.connect()
   })
 
   function createNUsightNetwork() {
@@ -58,7 +58,7 @@ describe('Networking Integration', () => {
       const onSensors = jest.fn()
       network.on(Sensors, onSensors)
 
-      robotSimulator.simulate()
+      virtualRobots.simulate()
 
       expect(onSensors).toHaveBeenCalledWith(expect.any(Sensors))
       expect(onSensors).toHaveBeenCalledTimes(1)
@@ -72,7 +72,7 @@ describe('Networking Integration', () => {
 
       network.off()
 
-      robotSimulator.simulate()
+      virtualRobots.simulate()
 
       expect(onSensors1).not.toHaveBeenCalled()
       expect(onSensors2).not.toHaveBeenCalled()
@@ -86,7 +86,7 @@ describe('Networking Integration', () => {
 
       off1()
 
-      robotSimulator.simulate()
+      virtualRobots.simulate()
 
       expect(onSensors1).not.toHaveBeenCalled()
       expect(onSensors2).toHaveBeenCalledWith(expect.any(Sensors))
@@ -108,7 +108,7 @@ describe('Networking Integration', () => {
 
       nusightNetwork.connect({ name: 'nusight' })
 
-      robotSimulator.simulate()
+      virtualRobots.simulate()
 
       expect(onSensors).toHaveBeenCalledWith(expect.any(Sensors))
     })
@@ -124,7 +124,7 @@ describe('Networking Integration', () => {
       const onSensors2 = jest.fn()
       network2.on(Sensors, onSensors2)
 
-      robotSimulator.simulate()
+      virtualRobots.simulate()
 
       expect(onSensors1).toHaveBeenCalledWith(expect.any(Sensors))
       expect(onSensors2).toHaveBeenCalledWith(expect.any(Sensors))
@@ -146,7 +146,7 @@ describe('Networking Integration', () => {
       const onSensors = jest.fn()
       localisationNetwork.on(Sensors, onSensors)
 
-      robotSimulator.simulate()
+      virtualRobots.simulate()
 
       expect(onSensors).toHaveBeenCalledTimes(1)
 
@@ -155,7 +155,7 @@ describe('Networking Integration', () => {
       const onVisionObject = jest.fn()
       visionNetwork.on(VisionObject, onVisionObject)
 
-      robotSimulator.simulate()
+      virtualRobots.simulate()
 
       expect(onVisionObject).toHaveBeenCalledTimes(0)
       expect(onSensors).toHaveBeenCalledTimes(1)
